chore: remove AgentEmbedSessionPage and its route
The /embed/session page is no longer used — the postMessage bootstrap
now lives directly in AgentEmbedPage. Remove the page, its test file,
the lazy import, and the route.
